+++ energyquantified/api/instances.py
@@ -104,6 +104,7 @@ class InstancesAPI(BaseAPI):
             frequency=None,
             aggregation=None,
             hour_filter=None,
+            threshold=None,
             ensembles=False):
         """
         Load time series instances.
@@ -146,7 +147,11 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param ensembles: Whether or not to include ensembles where available,\
+        :param threshold: Allow that many values to be missing within one frame of \
+            *frequency*. Has no effect unless *frequency* is provided, \
+            defaults to 0.
+        :type threshold: int, optional
+        :param ensembles: Whether to include ensembles where available,\
             defaults to False
         :type ensembles: bool, optional
         :return: A list :py:class:`energyquantified.data.Timeseries` instances
@@ -173,6 +178,7 @@ class InstancesAPI(BaseAPI):
         if "frequency" in params:
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
+            self._add_int(params, "threshold", threshold, min=0)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries_list(response.json())
@@ -186,6 +192,7 @@ class InstancesAPI(BaseAPI):
             frequency=None,
             aggregation=None,
             hour_filter=None,
+            threshold=None,
             ensembles=False):
         """
         Get the latest time series instance with filtering on `tags` and
@@ -214,7 +221,11 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param ensembles: Whether or not to include ensembles where available,\
+        :param threshold: Allow that many values to be missing within one frame of \
+            *frequency*. Has no effect unless *frequency* is provided, \
+            defaults to 0.
+        :type threshold: int, optional
+        :param ensembles: Whether to include ensembles where available,\
             defaults to False
         :type ensembles: bool, optional
         :return: A time series instance
@@ -236,6 +247,7 @@ class InstancesAPI(BaseAPI):
         if "frequency" in params:
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
+            self._add_int(params, "threshold", threshold, min=0)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries(response.json())
@@ -249,6 +261,7 @@ class InstancesAPI(BaseAPI):
             frequency=None,
             aggregation=None,
             hour_filter=None,
+            threshold=None,
             ensembles=False):
         """
         Get an instance specified by a `issued` (issue date) and `tag`. The
@@ -277,7 +290,11 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param ensembles: Whether or not to include ensembles where available,\
+        :param threshold: Allow that many values to be missing within one frame of \
+            *frequency*. Has no effect unless *frequency* is provided, \
+            defaults to 0.
+        :type threshold: int, optional
+        :param ensembles: Whether to include ensembles where available,\
             defaults to False
         :type ensembles: bool, optional
         :return: A time series instance
@@ -299,6 +316,7 @@ class InstancesAPI(BaseAPI):
         if "frequency" in params:
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
+            self._add_int(params, "threshold", threshold, min=0)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries(response.json())
@@ -317,7 +335,8 @@ class InstancesAPI(BaseAPI):
             time_zone=None,
             frequency=None,
             aggregation=None,
-            hour_filter=None):
+            hour_filter=None,
+            threshold=None):
         """
         Load one instance for each day based on some common criteria, stitch
         them together and return a continuous time series.
@@ -367,6 +386,10 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
+        :param threshold: Allow that many values to be missing within one frame of \
+            *frequency*. Has no effect unless *frequency* is provided, \
+            defaults to 0.
+        :type threshold: int, optional
         :return: A time series
         :rtype: :py:class:`energyquantified.data.Timeseries`
         """
@@ -390,6 +413,7 @@ class InstancesAPI(BaseAPI):
         if "frequency" in params:
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
+            self._add_int(params, "threshold", threshold, min=0)
         # Additional validation checks
         if sum(1 if t is not None else 0 for t in
                (time_of_day, after_time_of_day, before_time_of_day)) > 1: