+++ CHANGELOG.rst
@@ -2,11 +2,12 @@ Changelog
 =========
 
 dev
-------
+---
 
 **Improvements**
 
 - Add timezone conversion.
+- Add ``threshold`` parameter to define how many values are allowed to be missing while performing an aggregation
 
 0.10.1
 ------